refactor(mcp): split browser tool action dispatch for the file-size gate
The workflow_run/cancel actions pushed server/mcp/src/tools/browser.ts
past its 336-line baseline. Extract the per-action branch table into
browserActions.ts (runBrowserAction); browser.ts keeps the schema and
registration. Behaviour unchanged - the sidecar tool suite is green
across the move - and browser.ts drops to 187 lines, so its stale
file-size baseline entry is pruned (the gate ratchets down only).
