+++ scripts/file-size-baseline.json
@@ -1,11 +1,10 @@
 {
-  "_comment": "Frozen baseline of files exceeding the ~300-line rule (initial audit 20260612). scripts/check-file-size.mjs fails CI on NEW violators or growth of these. Burn entries down (split the file, lower or remove its number); never raise a number for logic growth. Rust entries were re-baselined 2026-07-02 to post-rustfmt line counts when `cargo fmt` became a CI gate — formatting normalization, zero logic change (see commit for the clippy/fmt pass); the ratchet direction still applies from these new values.",
-  "_note_browser_fmt_20260716": "browser/ai_commands.rs (302) and browser/origin_guard.rs (303) crossed 300 purely from `cargo fmt` line-wrapping when the fmt check was added to the pre-push gate (they were <300 pre-fmt; ai_commands.rs also took a mechanical clippy manual_map fix that removed lines). Formatting normalization, zero logic change — same case as the 2026-07-02 re-baseline noted above. Ratchet down from here (split or trim), never raise.",
+  "_comment": "Frozen baseline of files exceeding the ~300-line rule (initial audit 20260612). scripts/check-file-size.mjs fails CI on NEW violators or growth of these. Burn entries down (split the file, lower or remove its number); never raise a number for logic growth. Rust entries were re-baselined 2026-07-02 to post-rustfmt line counts when `cargo fmt` became a CI gate \u2014 formatting normalization, zero logic change (see commit for the clippy/fmt pass); the ratchet direction still applies from these new values.",
+  "_note_browser_fmt_20260716": "browser/ai_commands.rs (302) and browser/origin_guard.rs (303) crossed 300 purely from `cargo fmt` line-wrapping when the fmt check was added to the pre-push gate (they were <300 pre-fmt; ai_commands.rs also took a mechanical clippy manual_map fix that removed lines). Formatting normalization, zero logic change \u2014 same case as the 2026-07-02 re-baseline noted above. Ratchet down from here (split or trim), never raise.",
   "_note_embedded_browser_20260712": "WI-1.1: workspace.rs baseline ratcheted 632->385 by splitting its test module out to workspace.test.rs (a pure move; *.test.rs is excluded from the size count, so it is not baselined). WI-1.2: the Phase-0 spike_embed.rs probe was retired (recipe preserved in git cd162e02) and its baseline entry removed; lib.rs ratcheted 308->301 as those spike registrations left it, then grows ~1 line per registered browser command (WI-1.2/2.1/1.4 surface + driver + freeze/thaw = 310; WI-1.7 back/forward history = 312; WI-1.7 stop = 313; WI-1.7 confirm-dialog respond = 314). lib.rs is the app composition root (a command registry); this is registration growth, not logic bloat. surface.rs's macOS impl was split to surface_macos.rs (#[path]) to keep both under the 300 limit; the nav/UI delegate (WI-1.7/1.8) is a further #[path] split (nav_delegate_macos.rs); the run-loop pump/drive_load helpers split to driver_loop_macos.rs so surface_macos.rs stays under 300 as native rows are added.",
-  "_note_mcp_sidecar_20260728": "WI-7 (MCP audit remediation): scripts/check-file-size.mjs ROOTS gained server/mcp/src, which the gate had never scanned (audit 20260728 §4). Four entries froze the sidecar files that already exceeded 300 at their PRE-CHANGE sizes. Three have since ratcheted: server.ts left the baseline entirely (302 -> 160, after the eight unused arg extractors and the resource registry went), cli.ts 322 -> 317 (the dead resource-registration loop, minus the VERSION/capabilities comments that replaced it), and bridge/websocket.ts left it too (831 -> 299) when the audit's named seams were extracted into siblings — websocketConfig, connection, authHandshake, reconnect, pendingRequests, requestQueue, rateLimiter, wsProtocol — a pure refactor with the test file untouched. cli.ts then left the baseline too (317 -> 269) when detectClientIdentity moved to the new utils/clientIdentity.ts alongside the per-client credential reader — the extraction this note had already named. No sidecar file is baselined now. Never raise.",
+  "_note_mcp_sidecar_20260728": "WI-7 (MCP audit remediation): scripts/check-file-size.mjs ROOTS gained server/mcp/src, which the gate had never scanned (audit 20260728 \u00a74). Four entries froze the sidecar files that already exceeded 300 at their PRE-CHANGE sizes. Three have since ratcheted: server.ts left the baseline entirely (302 -> 160, after the eight unused arg extractors and the resource registry went), cli.ts 322 -> 317 (the dead resource-registration loop, minus the VERSION/capabilities comments that replaced it), and bridge/websocket.ts left it too (831 -> 299) when the audit's named seams were extracted into siblings \u2014 websocketConfig, connection, authHandshake, reconnect, pendingRequests, requestQueue, rateLimiter, wsProtocol \u2014 a pure refactor with the test file untouched. cli.ts then left the baseline too (317 -> 269) when detectClientIdentity moved to the new utils/clientIdentity.ts alongside the per-client credential reader \u2014 the extraction this note had already named. No sidecar file is baselined now. Never raise.",
   "files": {
     "server/content/src/server/createServer.ts": 411,
-    "server/mcp/src/tools/browser.ts": 336,
     "src-tauri/src/ai_provider/detection.rs": 370,
     "src-tauri/src/ai_provider/rest_api.rs": 403,
     "src-tauri/src/ai_provider/rest_providers.rs": 395,
